refactor: move gtag out of pages
